bugfix: Fix issues with latest mill version
Turns out latest versions show a synthetic target that has workspace/src path, which is taken instead of workspace/src/main/scala. This fixes it to take the longest prefix.
